+++ packages/web3-eth/CHANGELOG.md
@@ -263,3 +263,7 @@ Documentation:
 -   Fixed geth issue when running a new instance, transactions will index when there are no blocks created (#7098)
 
 ## [Unreleased]
+
+### Fixed
+
+-   Adds transaction property to be an empty list rather than undefined when no transactions are included in the block (#7151)